flowcount: rename plugin. fcount -> flowcount
