--- plugins/out_fcount/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(src
-  out_fcount.c)
-
-FLB_PLUGIN(out_fcount "${src}" "")
