+++ plugins/out_flowcount/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(src
+  out_flowcount.c)
+
+FLB_PLUGIN(out_flowcount "${src}" "")
